% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_tools.R
\name{change_proj_name}
\alias{change_proj_name}
\title{Change the project name}
\usage{
change_proj_name(path = "C:/", current_name, new_name)
}
\arguments{
\item{path}{the directory which is storing the project}

\item{current_name}{the current name of the project}

\item{new_name}{the new name of the directory}
}
\description{
This should only be done with extreme caution. Generally, the path should be C:/ drive
Make sure that you do not include uvp5_sn209 on the name path. 
This will recursively check all folders where the metadata are stored.
This operation may be unstable if done on a project which has been processed
Note I wrote this specifically for sn209. if using a different UVP you'd need to change it
}
