% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_config_file.R
\name{check_config_file}
\alias{check_config_file}
\title{Check the configuration files of a project}
\usage{
check_config_file(path = "C:/", proj_name)
}
\arguments{
\item{path}{the directory with the file lives}

\item{proj_name}{the name of a project, not including uvp5_sn209}
}
\description{
This function calls from reticulate. must have python 3 install & in path
you can set up reticulate with reticulate::py_config()
}
