% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_faulty_casts.R
\name{find_faulty}
\alias{find_faulty}
\title{Check for faulty casts}
\usage{
find_faulty(path = "C:/", proj_name)
}
\arguments{
\item{path}{directory holding project}

\item{proj_name}{project name}
}
\description{
This funciton will run through the raw casts and check for faulty casts
Sometimes casts will happen with no jpg's. These should be removed prior to metadata entry
Although they can also be removed after-the-fact
}
