% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bios_event_log.R
\name{grab_one}
\alias{grab_one}
\title{grab_two}
\usage{
grab_one(index, vect)
}
\arguments{
\item{index}{the starting point}

\item{vect}{the vector to pull from}
}
\description{
get next two lines in a vector after index
}
