% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_ctd.R
\name{min_dates}
\alias{min_dates}
\title{find min distance within 30mins}
\usage{
min_dates(date, ref_dates)
}
\arguments{
\item{date}{date}

\item{ref_dates}{ctd dates}
}
\description{
find min distance within 30mins
}
