% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_times.R
\name{proj_run_time}
\alias{proj_run_time}
\title{Get project run time}
\usage{
proj_run_time(raw_fnames = raw_fnames)
}
\arguments{
\item{raw_fnames}{the raw files names}
}
\description{
Get project run time
}
