% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bios_event_log.R
\name{read_bios_event}
\alias{read_bios_event}
\title{Read bios event files}
\usage{
read_bios_event(prog_id, prog_code)
}
\arguments{
\item{prog_id}{gf, hs, gfb, bv}

\item{prog_code}{number}
}
\description{
These can read in bios event logs
}
