% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_ctd.R
\name{recommend_ctd}
\alias{recommend_ctd}
\title{Recommend nearest CTD cast}
\usage{
recommend_ctd(event_log, raw_fnames)
}
\arguments{
\item{event_log}{a ctd event log from bios}

\item{raw_fnames}{the raw files}
}
\description{
Recommend nearest CTD cast
}
