% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_faulty_casts.R
\name{rem_faulty}
\alias{rem_faulty}
\title{Remove faulty casts}
\usage{
rem_faulty(path = "C:/", proj_name, bad_casts = bad_casts)
}
\arguments{
\item{path}{directory holding project}

\item{proj_name}{project name}

\item{bad_casts}{character of bad casts}
}
\description{
This should only be done after checking for faulties
}
