% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_tools.R
\name{write_admin_log}
\alias{write_admin_log}
\title{write an admin_log files}
\usage{
write_admin_log(proj_path, note)
}
\arguments{
\item{proj_path}{the path for the project}

\item{note}{a string of the note to create}
}
\description{
write an admin_log files
}
